% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{fit_em}
\alias{fit_em}
\title{A Cat Function}
\usage{
fit_em(
  long_labels,
  model = "BACE",
  max_iter = 1000,
  tol = 1e-08,
  verbose = FALSE
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
